+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/feat/SetupDetekt.kt
@@ -14,10 +14,10 @@ import fluxo.conf.impl.disableTask
 import fluxo.conf.impl.e
 import fluxo.conf.impl.l
 import fluxo.conf.impl.namedCompat
-import fluxo.conf.impl.onLibrary
 import fluxo.conf.impl.register
 import fluxo.conf.impl.splitCamelCase
 import fluxo.conf.impl.withType
+import fluxo.vc.onLibrary
 import io.github.detekt.gradle.DetektKotlinCompilerPlugin
 import io.gitlab.arturbosch.detekt.Detekt
 import io.gitlab.arturbosch.detekt.DetektCreateBaselineTask
@@ -214,7 +214,7 @@ internal fun Project.setupDetekt(
             input.from(detektTasks.map { it.sarifReportFile })
         }
 
-        context.libs?.run {
+        context.libs.run {
             dependencies {
                 onLibrary("detekt-formatting") { detektPlugins(it) }